requirements.txt
annotated-types==0.5.0
anyio==3.7.0
asgiref==3.7.2
certifi==2023.5.7
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pydantic==2.0
pydantic_core==2.0.1
pytest==7.4.0
pytest-django==4.5.2
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.0
